+++ includes/admin/settings/class.llms.settings.notifications.php
@@ -22,7 +22,9 @@ class LLMS_Settings_Notifications extends LLMS_Settings_Page {
 
 		add_filter( 'lifterlms_settings_tabs_array', array( $this, 'add_settings_page' ), 20 );
 		add_action( 'lifterlms_settings_' . $this->id, array( $this, 'output' ) );
+		add_action( 'lifterlms_settings_save_' . $this->id, array( $this, 'before_save' ), 5 );
 		add_action( 'lifterlms_settings_save_' . $this->id, array( $this, 'save' ) );
+		add_action( 'lifterlms_settings_save_' . $this->id, array( $this, 'after_save' ), 15 );
 		add_filter( 'llms_settings_' . $this->id . '_has_save_button', array( $this, 'maybe_disable_save' ) );
 
 	}
@@ -47,7 +49,7 @@ class LLMS_Settings_Notifications extends LLMS_Settings_Page {
 	 * @param    obj     $controller  instance of an LLMS_Notification_Controller
 	 * @return   array
 	 * @since    3.8.0
-	 * @version  3.8.0
+	 * @version  [version]
 	 */
 	private function get_notification_settings( $controller ) {
 
@@ -99,6 +101,13 @@ class LLMS_Settings_Notifications extends LLMS_Settings_Page {
 					'id' => $controller->get_option_name( $type . '_custom_subscribers' ),
 					'type' => 'text',
 				);
+			}
+		}
+
+		if ( $controller->is_testable( $type ) ) {
+			foreach ( $controller->get_test_settings( $type ) as $setting ) {
+				$setting['id'] = 'llms_notification_test_data[' . $setting['id'] . ']';
+				$settings[] = $setting;
 			}
 		}
 
@@ -186,6 +195,46 @@ class LLMS_Settings_Notifications extends LLMS_Settings_Page {
 	public function merge_code_button() {
 
 		llms_merge_code_button( $this->view->get_option_name( 'body' ), true, $this->view->get_merge_codes() );
+
+	}
+
+	/**
+	 * Remove test data from $_POST so that it wont be saved to the DB
+	 * @return   void
+	 * @since    [version]
+	 * @version  [version]
+	 */
+	public function before_save()	 {
+
+		if ( isset( $_POST['llms_notification_test_data'] ) ) {
+
+			$_POST['llms_notification_test_data_temp'] = $_POST['llms_notification_test_data'];
+			unset( $_POST['llms_notification_test_data'] );
+
+		}
+
+	}
+
+	/**
+	 * Send a test notification after notification data is saved
+	 * @return   void
+	 * @since    [version]
+	 * @version  [version]
+	 */
+	public function after_save() {
+
+		if ( isset( $_GET['notification'] ) && isset( $_GET['type'] ) && isset( $_POST['llms_notification_test_data_temp'] ) ) {
+
+			if ( ! empty( $_POST['llms_notification_test_data_temp']['submission_id'] ) ) {
+
+				$submission = llms_get_assignment_submission( $_POST['llms_notification_test_data_temp']['submission_id'] );
+				$controller = LLMS()->notifications()->get_controller( $_GET['notification'] );
+				$controller->send_test( sanitize_text_field( $_GET['type'] ), $submission );
+
+			}
+
+		}
+
 
 	}